instruct_gs/requirements.txt
# InstructGS Requirements
# Dependencies for text-to-3D editing with Gaussian Splatting

# Core dependencies (should be installed with nerfstudio)
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Diffusion model dependencies
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0

# Image processing
Pillow>=9.0.0

# Optional: for better performance
# xformers>=0.0.20  # Uncomment for memory-efficient attention

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0